V.1.3.1

+++ build/build-tools.js
@@ -2,15 +2,16 @@ const fs = require('fs');
 const path = require('path');
 
 const copy = (src, dest) => {
-    console.log(`Copying from ${src} to ${dest}`);
-    
-    fs.cp(src, dest, {
-        recursive: true
-    }, (err) => {
-        if (err) {
-            console.log(err);
-        }
-    });
+    if (fs.existsSync(src)) {
+        fs.cp(src, dest, {
+            recursive: true
+        }, (err) => {
+            if (err) {
+                console.log(err);
+            }
+        });
+        fs.existsSync
+    }
 };
 
 module.exports = { copy };
